+++ src/main/kotlin/kscript/app/creator/IdeaProjectCreator.kt
@@ -8,6 +8,8 @@ import kscript.app.util.Logger.devMsg
 import kscript.app.util.Logger.infoMsg
 import java.net.URI
 import java.nio.file.Path
+import kotlin.io.path.exists
+import kotlin.io.path.extension
 
 class IdeaProjectCreator {
     fun create(basePath: Path, script: Script, userArgs: List<String>, uriLocalPathProvider: (URI) -> Path): Path {
@@ -15,20 +17,30 @@ class IdeaProjectCreator {
 
         for (scriptNode in script.scriptNodes) {
             val sourceUri = scriptNode.sourceUri
-            val path = basePath.resolve("src/${scriptNode.scriptName}")
+            var path = basePath.resolve("src/${scriptNode.scriptName + scriptNode.scriptType.extension}")
 
             if (sourceUri == null) {
                 FileUtils.createFile(path, scriptNode.sections.joinToString("\n") { it.code })
             } else {
                 val targetPath = uriLocalPathProvider(sourceUri)
                 devMsg("link: $path, target: $targetPath")
+
+                var counter = 0
+
+                while (path.exists()) {
+                    //Duplicated script name e.g. from different sources
+                    //We will try to add extension until it is correct name
+                    path = basePath.resolve("src/${scriptNode.scriptName + "_$counter" + scriptNode.scriptType.extension}")
+                    counter++
+                }
+
                 FileUtils.symLinkOrCopy(path, targetPath)
             }
         }
 
         FileUtils.createFile(
             basePath.resolve(".idea/runConfigurations/Main.xml"),
-            Templates.runConfig(script.rootNode.scriptName, userArgs)
+            Templates.runConfig(script.rootNode.scriptName + script.rootNode.scriptType.extension, userArgs)
         )
 
         FileUtils.createFile(